fix: #38-제너릭 타입으로 변경
