+++ backend/ezgg/src/main/java/com/matching/ezgg/domain/matchInfo/matchKeyword/rule/KeywordRule.java
@@ -1,10 +1,8 @@
 package com.matching.ezgg.domain.matchInfo.matchKeyword.rule;
 
-import com.matching.ezgg.domain.matchInfo.matchKeyword.dto.GlobalMatchParsingDto;
-import com.matching.ezgg.domain.matchInfo.matchKeyword.keyword.GlobalKeyword;
 import com.matching.ezgg.domain.matchInfo.matchKeyword.lane.Lane;
 
-public interface KeywordRule {
-	Boolean matchWithRule(GlobalMatchParsingDto globalMatchParsingDto, Lane lane);
-	GlobalKeyword getKeyword();
+public interface KeywordRule<Dto, Keyword> {
+	Boolean matchWithRule(Dto parsingDto, Lane lane);
+	Keyword getKeyword();
 }
