+++ backend/ezgg/src/main/java/com/matching/ezgg/domain/matchInfo/matchKeyword/rule/globalRule/HighKdaRule.java
@@ -1,18 +1,19 @@
-package com.matching.ezgg.domain.matchInfo.matchKeyword.rule;
+package com.matching.ezgg.domain.matchInfo.matchKeyword.rule.globalRule;
 
 import com.matching.ezgg.domain.matchInfo.matchKeyword.dto.GlobalMatchParsingDto;
 import com.matching.ezgg.domain.matchInfo.matchKeyword.keyword.GlobalKeyword;
 import com.matching.ezgg.domain.matchInfo.matchKeyword.lane.Lane;
+import com.matching.ezgg.domain.matchInfo.matchKeyword.rule.KeywordRule;
 
-public class HighDamageRule implements KeywordRule {
+public class HighKdaRule implements KeywordRule<GlobalMatchParsingDto, GlobalKeyword> {
 	@Override
 	public Boolean matchWithRule(GlobalMatchParsingDto globalMatchParsingDto, Lane lane) {
 		Lane.Criteria criteria = lane.getCriteria();
-		return criteria!=null && globalMatchParsingDto.getDamagePerMinute() >= criteria.damagePerMinute();
+		return criteria!=null && globalMatchParsingDto.getKda() >= criteria.kda();
 	}
 
 	@Override
 	public GlobalKeyword getKeyword() {
-		return GlobalKeyword.HIGH_DAMAGE;
+		return GlobalKeyword.HIGH_KDA;
 	}
 }
