+++ backend/ezgg/src/main/java/com/matching/ezgg/domain/matchInfo/matchKeyword/rule/globalRule/LevelDiffRule.java
@@ -1,17 +1,18 @@
-package com.matching.ezgg.domain.matchInfo.matchKeyword.rule;
+package com.matching.ezgg.domain.matchInfo.matchKeyword.rule.globalRule;
 
 import com.matching.ezgg.domain.matchInfo.matchKeyword.dto.GlobalMatchParsingDto;
 import com.matching.ezgg.domain.matchInfo.matchKeyword.keyword.GlobalKeyword;
 import com.matching.ezgg.domain.matchInfo.matchKeyword.lane.Lane;
+import com.matching.ezgg.domain.matchInfo.matchKeyword.rule.KeywordRule;
 
-public class NeverGivesUpRule implements KeywordRule {
+public class LevelDiffRule implements KeywordRule<GlobalMatchParsingDto, GlobalKeyword> {
 	@Override
 	public Boolean matchWithRule(GlobalMatchParsingDto globalMatchParsingDto, Lane lane) {
-		return globalMatchParsingDto.getGameEndedInSurrender() == Boolean.FALSE;
+		return globalMatchParsingDto.getMaxLevelLeadLaneOpponent() >= 2;
 	}
 
 	@Override
 	public GlobalKeyword getKeyword() {
-		return GlobalKeyword.NEVER_GIVES_UP;
+		return GlobalKeyword.LEVEL_DIFF;
 	}
 }
